sidecar/requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
cryptography==46.0.3
python-dotenv==1.2.2
aiosqlite==0.21.0
SQLAlchemy==2.0.45
pydantic==2.12.5
